lz4: fix compress max output bound
Max output size can be larger than the source length if the input data
is 'too' random. This is normal and thus should be allowed since the
caller is expected to have the actual size of the output written to be
updated via the dst_len variable.

Bug: 111209481
Change-Id: I44a8abb6fb55936f2fd47616954341ba6b0d62ec
Signed-off-by: David Lin <dtwlin@google.com>
Signed-off-by: root <franciscofranco.1990@gmail.com>
(cherry picked from commit 46821588b36ab1ae07247fe3b6d245f9de88362f)